modified the gradle file and changed the id to Integer

+++ src/main/java/com/acoustic/repository/TotalZusRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TotalZusRepository extends CrudRepository<TotalZus, Long> {
+public interface TotalZusRepository extends CrudRepository<TotalZus, Integer> {
